refactor(tensor): remove Scalar bound from Tensor<T> struct (#554)
Move `T: Scalar` from the struct definition to individual impl blocks
that actually need arithmetic operations. This enables `Tensor<bool>`,
`Tensor<u8>`, `Tensor<i32>` etc. for pure storage/masking use cases.

## Bound hierarchy after change

- `impl<T> Tensor<T>` — metadata, view ops (permute, reshape, etc.)
- `impl<T: Conjugate> Tensor<T>` — mH (Hermitian transpose)
- `impl<T: Scalar> Tensor<T>` — constructors (zeros, ones, from_slice),
  contiguous/into_contiguous, cat/stack, tril/triu, Differentiable

Closes #554
